Change all offsets to usize
Normalizing on `usize` is more consistent and idiomatic. This
eliminates the need to cast from `usize` to `u32` or `u64`.

The `offset_of!` macro has also been removed and replaced with
the memoffset crate.

Added minor optimization for command buffer recording to reduce
allocations (use SmallVec instead of Vec for offsets and buffers).

+++ src/lib.rs
@@ -800,7 +800,7 @@ pub struct VertexAttributeDescriptor {
     pub shader_location: u32,
     /// The index of the vertex buffer containing this attribute. See `RenderPassEncoder::set_vertex_buffers`.
     pub input_slot: u32,
-    pub offset: u32,
+    pub offset: usize,
     pub format: VertexFormat,
 }
